Fixed many issues due to merge with master
