+++ shimming-toolbox/test/cli/test_cli_b0shim.py
@@ -1018,7 +1018,7 @@ class TestCLIRealtime(object):
                 lines = file.readlines()
                 line = lines[5].strip().split(',')
                 values = [float(val) for val in line if val.strip()]
-            assert values == [10.809849, -0.015250208108, 1454.1917]
+            assert np.allclose(values, [10.809849, -0.015250208108, 1454.1917])
 
     def test_cli_rt_time_offset_auto(self, nii_fmap, nii_target, nii_mask, fm_data, target_data):
         with tempfile.TemporaryDirectory(prefix='st_' + pathlib.Path(__file__).stem) as tmp:
@@ -1060,7 +1060,7 @@ class TestCLIRealtime(object):
                 lines = file.readlines()
                 line = lines[5].strip().split(',')
                 values = [float(val) for val in line if val.strip()]
-            assert values == [10.834912, -0.014710841563, 1504.1091]
+            assert values == [10.901898, -0.014273692604, 1537.7406]
 
     def test_cli_rt_sph_order_0(self, nii_fmap, nii_target, nii_mask, fm_data, target_data):
         with tempfile.TemporaryDirectory(prefix='st_' + pathlib.Path(__file__).stem) as tmp:
